feat :  connect EventInfoFree with backend database
