+++ src/components/events/EventInfoFree.vue
@@ -1,137 +1,92 @@
 <script setup>
-import { ref, computed, onMounted } from 'vue';
-import { useRoute } from 'vue-router'
-import dayjs from 'dayjs';
-import axios from 'axios'
-import 'dayjs/locale/zh-tw'
-dayjs.locale('zh-tw')
+import { useEvent } from '@/composable/useEvent.js';
+import { toRef } from 'vue';
+import EventHoster from './EventHoster.vue';
+import MessageBoard from './MessageBoard.vue';
 
-const isJoin = ref(false)
-const showModal = ref(false)
-const event = ref(null)
-const now = ref(dayjs())
-const route = useRoute()
-const eventId = route.params.id
-const notFound = ref(false)
-const isLoading = ref(true)
-const errorMsg = ref('')
+const props = defineProps({
+  event: Object
+}) 
 
+const eventRef = toRef(props, 'event')
 
-onMounted( async() => {
-  try{
-      const res = await axios.get(`/event/${eventId}`)
-      event.value = res.data
-      console.log(event.value)
-  }catch(err){
-    if( err.response && err.response.status == 404){
-      notFound.value = true
-      
-    }else{
-      errorMsg.value = '取得活動資料失敗'
-      console.error('取得活動資料失敗', err)
-    }
-  }finally{
-    isLoading.value = false
-    }
-})
+const {
+  isJoin,
+  toggleJoin,
+  isOver24hr,
+  showModal,
+  formattedEventTime,
+  openCancelModal,
+  closeModal,
+  handleConfirmCancel
+} = useEvent(eventRef)
 
-const isOver24hr = computed(() => {
-  if( !event.value || !event.value.startDate ) return false
-  return dayjs(event.value.startDate).diff(now.value, 'hour') > 24
-})
-
-const formattedEventTime = computed(() =>{
-  if (!event.value) return ''
-  const start = dayjs(event.value.startDate).format('YYYY.MM.DD HH:mm (ddd)')
-  const end = dayjs(event.value.endDate).format('YYYY.MM.DD HH:mm (ddd)')
-  return `${start} ~ ${end}`
-})
-
-function toggleJoin(){
-  isJoin.value = !isJoin.value
-  console.log('現在狀態是：', isJoin.value)
-}
-
-function openCancelModal(){
-  showModal.value = true
-}
-
-function closeModal() {
-  showModal.value = false
-}
-
-function handleConfirmCancel(){
-  isJoin.value = false
-  showModal.value = false
-}
 
 </script>
 
 <template>
+
   <div>
-    <p v-if="isLoading">載入中，請稍後...</p>
-    <p v-else-if="notFound">找不到活動</p>
-    <p v-else-if="errorMsg">{{ errorMsg }}</p>
-    <div v-else>
-      <div :class="['modal', { 'modal-open': showModal }]">
-        <div class="modal-box">
-          <h3 class="text-lg font-bold">確認取消報名</h3>
-          <p class="py-4">
-            您確定要取消這次報名嗎？ <br/>
-            <span>取消後如人數額滿或是活動開始前24小時內都將無法報名</span>， <br/>
-            請再次確認您的選擇。
-          </p>
-          <div class="modal-action">
-            <button class="btn" @click="closeModal">放棄取消</button>
-            <button class="btn" @click="handleConfirmCancel">確認取消</button>
-          
+    <div :class="['modal', { 'modal-open': showModal }]">
+      <div class="modal-box">
+        <h3 class="text-lg font-bold">確認取消報名</h3>
+        <p class="py-4">
+          您確定要取消這次報名嗎？ <br/>
+          <span>取消後如人數額滿或是活動開始前24小時內都將無法報名</span>， <br/>
+          請再次確認您的選擇。
+        </p>
+        <div class="modal-action">
+          <button class="btn" @click="closeModal">放棄取消</button>
+          <button class="btn" @click="handleConfirmCancel">確認取消</button>
+        
+        </div>
+      </div>
+    </div>
+  
+      <div class="event-information-section">
+        <div class="event-information-card">
+          <div class="event-img">
+            <img src="@/components/events/picture/酒吧示意圖.jpg" alt="酒吧示意圖"> 
+          </div>
+          <div class="event-content-box">
+            <div class="event-map">
+            </div>
+            <div class="event-content">
+              <div class="event-tags">
+                <div>免費活動</div>
+                <div>下班來喝</div>
+              </div>
+              <div>
+                <h3 class="event-title">{{ props.event.name }}
+                </h3>
+                <div class="event-content-info">
+                  <i class="fa-solid fa-calendar"></i>
+                  <p v-if="formattedEventTime">活動時間：{{ formattedEventTime }}</p>
+                </div>
+                <div class="event-content-info">
+                  <i class="fa-solid fa-wine-glass"></i>
+                
+                  <p>店名：{{ props.event.barName}}</p>
+                </div>
+                <div class="event-content-info">
+                  <i class="fa-solid fa-location-dot"></i>
+                  <p>地址：{{ props.event.location }}</p>
+                </div>
+                <div class="event-content-info">
+                  <i class="fa-solid fa-user"></i>
+                  <p>目前報名人數： <span>12</span> ｜ 報名人數上限：<span>{{ props.event.maxPeople || '無報名人數限制' }}</span></p>
+                </div>
+              </div>
+              <button @click="toggleJoin()" :disabled = "isJoin" :class="{ 'opacity-50 cursor-not-allowed': isJoin }" type="button" class="event-btn event-btn-free" >{{ isJoin ? '已報名' : '參加活動' }}</button>
+              <button v-if="isJoin" @click="openCancelModal()" :disabled="!isOver24hr" :class="['event-btn event-btn-free',    isOver24hr ? 'cursor-pointer' : 'cursor-not-allowed opacity-50']"  type="button" class="event-btn event-btn-free" >取消報名</button>
+            
+            </div>
           </div>
         </div>
       </div>
-    
-        <div class="event-information-section">
-          <div class="event-information-card">
-            <div class="event-img">
-              <img src="@/components/events/picture/酒吧示意圖.jpg" alt="酒吧示意圖"> 
-            </div>
-            <div class="event-content-box">
-              <div class="event-map">
-              </div>
-              <div class="event-content">
-                <div class="event-tags">
-                  <div>免費活動</div>
-                  <div>下班來喝</div>
-                </div>
-                <div>
-                  <h3 class="event-title">{{ event.name }}
-                  </h3>
-                  <div class="event-content-info">
-                    <i class="fa-solid fa-calendar"></i>
-                    <p v-if="formattedEventTime">活動時間：{{ formattedEventTime }}</p>
-                  </div>
-                  <div class="event-content-info">
-                    <i class="fa-solid fa-wine-glass"></i>
-                  
-                    <p>店名：{{ event.barName}}</p>
-                  </div>
-                  <div class="event-content-info">
-                    <i class="fa-solid fa-location-dot"></i>
-                    <p>地址：{{ event.location }}</p>
-                  </div>
-                  <div class="event-content-info">
-                    <i class="fa-solid fa-user"></i>
-                    <p>目前報名人數： <span>12</span> ｜ 報名人數上限：<span>{{ event.maxPeople || '無報名人數限制' }}</span></p>
-                  </div>
-                </div>
-                <button @click="toggleJoin()" :disabled = "isJoin" :class="{ 'opacity-50 cursor-not-allowed': isJoin }" type="button" class="event-btn event-btn-free" >{{ isJoin ? '已報名' : '參加活動' }}</button>
-                <button v-if="isJoin" @click="openCancelModal()" :disabled="!isOver24hr" :class="['event-btn event-btn-free',    isOver24hr ? 'cursor-pointer' : 'cursor-not-allowed opacity-50']"  type="button" class="event-btn event-btn-free" >取消報名</button>
-              
-              </div>
-            </div>
-          </div>
-        </div>
-    </div>  
-  </div>
+      <!-- <EventHoster /> -->
+      <!-- <MessageBoard v-if="isJoin" /> -->
+  </div>  
 </template>
 
 <style scoped>
@@ -143,8 +98,6 @@ function handleConfirmCancel(){
   justify-content: center;
   align-items: center;
 }
-
-
 
 .event-information-card{
   max-width: 1200px;
