+++ src/views/events/EventInfo.vue
@@ -1,28 +1,64 @@
 <script setup>
+import { useRoute } from 'vue-router'
+import { ref, onMounted, computed } from 'vue';
+import axios from 'axios'
+
 import EventInfoFree from '@/components/events/EventInfoFree.vue'
 import EventInfoPay from '@/components/events/EventInfoPay.vue'
-import EventHoster from '@/components/events/EventHoster.vue'
-import MessageBoard from '@/components/events/MessageBoard.vue'
+
+const route = useRoute()
+const eventId = route.params.id
+const isLoading = ref(true)
+const errorMsg = ref('')
+const event = ref(null)
+const notFound = ref(false)
+
+// 設定axios
 
 
+onMounted( async() => {
+  try{
+      const res = await axios.get(`/api/event/${eventId}`)
+      const { stringModel, tagIds } = res.data
 
+      event.value = stringModel
+      // eventTagIds.value = tagIds
 
+      console.log('活動資料:', stringModel)
+      console.log('活動標籤 IDs:', tagIds)
+      // console.log(`=========================${res.data.stringModel}`)
+  }catch(err){
+    if( err.response && err.response.status == 404){
+      notFound.value = true
+    }else{
+      errorMsg.value = '取得活動資料失敗'
+      console.error('取得活動資料失敗', err)
+    }
+  }finally{
+    isLoading.value = false
+    }
+})
+
+const isFree = computed(() => {
+  return event.value?.price == "" || event.value?.price == null || event.value?.price == undefined
+})
 
 </script>
 
 <template>
+  <div>
+    <p v-if="isLoading">載入中，請稍後...</p>
+    <p v-else-if="notFound">找不到活動</p>
+    <p v-else-if="errorMsg">{{ errorMsg }}</p>
 
-  <div class="page">
-    <h1>酒吧活動</h1>
-    <p>參加近期活動，結識朋友！</p>
-    <p>探索附近的酒吧活動。</p>
+    <EventInfoFree v-else-if="isFree" :event="event" />
+    <EventInfoPay v-else :event="event" />
+
+
   </div>
 
 
-  <EventInfoFree />
-  <EventInfoPay />
-  <EventHoster />
-  <MessageBoard />
+
 
 </template>
