+++ vite.config.js
@@ -1,18 +1,26 @@
 import { fileURLToPath, URL } from 'node:url'
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
-import path from 'path';
+import path from 'path'
 import vueDevTools from 'vite-plugin-vue-devtools'
-import tailwindcss from '@tailwindcss/vite';
+import tailwindcss from '@tailwindcss/vite'
 
-// https://vite.dev/config/
+
 export default defineConfig({
-  plugins: [vue(),
-    tailwindcss()
+  plugins: [
+    vue(),
+    tailwindcss(),
+    vueDevTools()
   ],
   resolve: {
     alias: {
       '@': path.resolve(__dirname, 'src'),
     }
+  },
+  server: {
+    proxy: {
+      '/api':'http://localhost:3000'
+      }
+    }
   }
-})
+)
